fix twitter

+++ twitter.py
@@ -1,18 +1,14 @@
 import logging
 import os
-from typing import Optional
+from typing import Optional, Union, Tuple
 
 import telegram
-import tweepy as tweepy
 from dotenv import load_dotenv
-from lxml import html
 from lxml.html import fromstring
-from telegram import Update
 from telegram.ext import CallbackContext
 from tweepy import Client, API, OAuth1UserHandler
 
 from data.db import Post
-from util.helper import get_caption, get_file
 
 load_dotenv()
 
@@ -34,7 +30,6 @@ api_DE = API(OAuth1UserHandler(
     access_token=access_token_DE,
     access_token_secret=access_secret_DE,
 ))
-
 
 consumer_key_EN = os.getenv("CONSUMER_KEY_EN")
 consumer_secret_EN = os.getenv("CONSUMER_SECRET_EN")
@@ -59,29 +54,18 @@ ACTIVE = True
 TWEET_LENGTH = 280
 
 
-def supply_twitter_instance(lang_key: Optional[str]=None):
-    def decorator(func):
-        async def wrapper(*args, **kwargs):
-            if not ACTIVE:
-                return
-
-            if lang_key == "en":
-                client = client_EN
-                api = api_EN
-            elif lang_key is None:
-                client = client_DE
-                api = api_DE
-            else:
-                return
-
-            return await func(client, api, *args, **kwargs)
-
-        return wrapper
-
-    return decorator
+def supply_twitter_instance(lang_key: Optional[str] = None) -> Union[Tuple[Client, API], None]:
+    if not ACTIVE:
+        return None
+    if lang_key == "en":
+        return client_EN, api_EN
+    elif lang_key is None:
+        return client_DE, api_DE
+    else:
+        return None
 
 
-def upload_media(files, api:API):
+def upload_media(files, api: API):
     media_ids = []
     for file in files:
         res = api.media_upload(file)
@@ -89,7 +73,7 @@ def upload_media(files, api:API):
     return media_ids
 
 
-def create_tweet(text:str, client:Client,media_ids=None,):
+def create_tweet(text: str, client: Client, media_ids=None, ):
     text = fromstring(text).text_content().strip()
     try:
         client.create_tweet(text=text.replace("\n", " ").replace("  ", " ")[:TWEET_LENGTH], media_ids=media_ids)
@@ -97,22 +81,36 @@ def create_tweet(text:str, client:Client,media_ids=None,):
         logging.error(f"Error when trying to post to twitter: {e}\n\ntext: {text}\n\nmedia_ids: {media_ids}")
 
 
-@supply_twitter_instance
-async def tweet_file(text: str, file: telegram.File, client:Client,api:API):
+async def tweet_file(text: str, file: telegram.File, lang_key: Optional[str] = None):
+    instance = supply_twitter_instance(lang_key)
+    if instance is None:
+        return
+    client, api = instance
+
     path = file.file_path.split('/')[-1]
     await file.download_to_drive(path)
-    media_ids = upload_media([path],api)
-    create_tweet(text,client, media_ids)
+    media_ids = upload_media([path], api)
+    create_tweet(text, client, media_ids)
     os.remove(path)
 
-@supply_twitter_instance
-async def tweet_file_3(text: str, path: str,api:API,client:Client):
+
+async def tweet_file_3(text: str, path: str, lang_key: Optional[str] = None):
+    instance = supply_twitter_instance(lang_key)
+    if instance is None:
+        return
+    client, api = instance
+
     media_id = api.media_upload(path)
-    create_tweet(text, client,[media_id.media_id])
+    create_tweet(text, client, [media_id.media_id])
 
 
 @supply_twitter_instance
-async def tweet_files(context: CallbackContext, text: str, client:Client,api:API, posts: [Post],):
+async def tweet_files(context: CallbackContext, text: str, posts: [Post], lang_key: Optional[str] = None):
+    instance = supply_twitter_instance(lang_key)
+    if instance is None:
+        return
+    client, api = instance
+
     upload_files = []
     for post in posts:
         file = await context.bot.get_file(post.file_id)
@@ -120,9 +118,7 @@ async def tweet_files(context: CallbackContext, text: str, client:Client,api:API
         await file.download_to_drive(path)
         upload_files.append(path)
 
-    media_ids = upload_media(upload_files,api)
+    media_ids = upload_media(upload_files, api)
     create_tweet(text, client, media_ids)
     for path in upload_files:
         os.remove(path)
-
-
